youtube_knowledge/pyproject.toml
[project]
name = "youtube-knowledge"
version = "0.1.0"
description = "Transform YouTube playlists into Gemini knowledge bases for semantic search and Q&A"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "youtube-transcript-api>=0.6.0",
    "yt-dlp>=2024.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pyright>=1.1.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
yt_knowledge = "youtube_knowledge.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
# Set the maximum line length to 100 characters
line-length = 100

# Target Python 3.11+
target-version = "py311"

# Exclude common directories
exclude = [
    ".git",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
# Enable specific rule sets
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "SIM",   # flake8-simplify
    "RET",   # flake8-return
    "ARG",   # flake8-unused-arguments
    "PTH",   # flake8-use-pathlib
    "PL",    # pylint
    "RUF",   # Ruff-specific rules
]

# Ignore specific rules that might be too strict for this project
ignore = [
    "E501",    # Line too long (handled by formatter)
    "PLR0913", # Too many arguments to function call
    "PLR2004", # Magic value used in comparison
    "RET504",  # Unnecessary variable assignment before return
    "SIM108",  # Use ternary operator (not always clearer)
    "ARG001",  # Unused function argument (common in callbacks)
    "ARG002",  # Unused method argument (common in overrides)
]

# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
# Ignore certain rules in CLI files that use click decorators
"youtube_knowledge/cli.py" = ["ARG001"]

[tool.ruff.lint.isort]
# Configure import sorting
known-first-party = ["youtube_knowledge"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Indent with spaces
indent-style = "space"

# Respect existing trailing commas
skip-magic-trailing-comma = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=youtube_knowledge",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
]

[tool.pyright]
# Python version and paths
pythonVersion = "3.11"
include = ["youtube_knowledge"]
exclude = [
    "**/__pycache__",
    ".venv",
    ".git",
    "build",
    "dist",
    "tests",  # Exclude tests from type checking for now
]

# Type checking mode
typeCheckingMode = "basic"  # Options: "off", "basic", "standard", "strict"

# Reporting rules
reportMissingImports = true
reportMissingTypeStubs = false
reportUnusedImport = true
reportUnusedClass = true
reportUnusedFunction = true
reportUnusedVariable = true
reportDuplicateImport = true
reportOptionalMemberAccess = true
reportOptionalSubscript = true
reportOptionalCall = true
reportOptionalIterable = true
reportUntypedFunctionDecorator = false  # Click decorators
reportUntypedClassDecorator = false
reportUntypedBaseClass = true
reportUnknownParameterType = false  # Too strict for basic mode
reportUnknownArgumentType = false  # Too strict for basic mode
reportUnknownVariableType = false  # Too strict for basic mode
reportUnknownMemberType = false  # Too strict for basic mode
reportMissingTypeArgument = false
reportPrivateUsage = true
reportConstantRedefinition = true
reportIncompatibleMethodOverride = true
reportIncompatibleVariableOverride = true
reportInconsistentConstructor = true
